+++ NEWS.md
@@ -3,6 +3,7 @@
 ### Minor improvement
 
 - `update(, collapse_m = m)` will now produce a level "other p" for the p least frequent categories [#52](https://github.com/mayer79/effectplots/pull/52).
+- `update()` has received new defaults: `collapse_m = 15` (was 30), `drop_empty = TRUE` (was `FALSE`) [#54](https://github.com/mayer79/effectplots/pull/54).
 
 ### Minor changes